Minor Fixes

+++ lib/include/factory/neo4j_interface.h
@@ -97,7 +97,7 @@ public:
   virtual int get_int_element(std::string key) = 0;
 
   //! Get a float element out of a map
-  virtual float get_float_element(std::string key) = 0;
+  virtual double get_float_element(std::string key) = 0;
 
   //! Get a map element out of a map
   virtual DbMapInterface* get_map_element(std::string key) = 0;
